@@ -22,7 +22,7 @@
     "url": "git://github.com/railnerd/nce.git"
   },
   "engines": {
-    "node": ">= 0.8.0"
+    "node": ">= 6.0.0"
   },
   "license": "MIT",
   "readmeFilename": "README.md"
